DA mappers and dtos expanded and enhanced
