+++ src/SklaDinya_desktop/SklaDinya_desktop-DA-component/Dtos/AuthDtos.cs
@@ -1,0 +1,11 @@
+namespace SklaDinya_desktop_DA_component.Dtos;
+
+internal record LoginRequest(
+    string Username,
+    string Password);
+
+internal record RegistrationRequest(
+    string  Username,
+    string  Password,
+    string  Name,
+    string? Email);
