+++ src/SklaDinya_desktop/SklaDinya_desktop-DA-component/Dtos/PaymentDtos.cs
@@ -1,0 +1,3 @@
+namespace SklaDinya_desktop_DA_component.Dtos;
+
+internal record PaymentRequest(string Receipt);
