+++ src/SklaDinya_desktop/SklaDinya_desktop-DA-component/Dtos/PriceDtos.cs
@@ -1,0 +1,11 @@
+namespace SklaDinya_desktop_DA_component.Dtos;
+
+internal record PriceDto(
+    string? StorageId,
+    string? CellClass,
+    string? Price,
+    string? CreatedAt);
+
+internal record PriceCreateRequest(
+    string CellClass,
+    string Price);
